Move not found page into component
