+++ app/workouts/edit/page.tsx
@@ -1,8 +1,9 @@
 import Link from "next/link";
 import { auth } from "@clerk/nextjs";
 import { db } from "@/db/drizzle";
+import WorkoutNotFound from "../WorkoutNotFound";
 import { Workout, TWorkoutFormSchema } from "@/lib/types";
-import WorkoutForm from "@/components/WorkoutForm";
+import WorkoutForm from "@/app/workouts/WorkoutForm";
 
 async function getWorkout(id: number) {
   // TODO: We do want to enforce userId so people can't look at other users' workouts
@@ -50,20 +51,7 @@ export default async function EditWorkoutPage({
   const workout = await getWorkout(workoutId);
 
   if (!workout) {
-    return (
-      // TODO: Put this into a separate component
-      <div className="mx-auto flex flex-col p-4 text-center sm:max-w-md">
-        <p className="text-left">
-          Workout not found, please return to the Workouts page
-        </p>
-        <Link
-          className="mx-auto mt-2 rounded-md bg-slate-500 p-2 text-white hover:bg-slate-500/70"
-          href="/workouts"
-        >
-          Workouts
-        </Link>
-      </div>
-    );
+    return <WorkoutNotFound></WorkoutNotFound>;
   } else {
     return (
       <WorkoutForm
